Added code to be able to control the bin arm, the front elevator, and the rear elevator in autonomous mode.
Added code to be able to control the bin arm, the front elevator, and
the rear elevator in autonomous mode.  This included new commands
AutonDropBinArm, AutonDropFront, AutonDropRear, AutonLiftBinArm,
AutonLiftRear.  Also fixed AutonLiftFront to utilize the front elevator
as it had a bug and was using the rear elevator.  This was not a problem
in the past because we never used the front elevator in autonomous mode.